Feature/#91: CalendarHeader 일부 수정
