+++ Harubee-iOS/Sources/Presentation/Calendar/CalendarView.swift
@@ -9,7 +9,7 @@
 import SwiftUI
 import DesignSystem
 
-// MARK: - 캘린더 뷰 설명
+// MARK: - Calendar View Description
 ///
 /// `CalendarView`는 월급 기반 예산 관리 앱의 핵심 캘린더 인터페이스를 구현하는 View입니다.
 /// 이 뷰는 사용자의 월급일을 기준으로 설정된 예산 기간을 캘린더 형식으로 표시하며,
@@ -23,7 +23,7 @@ import DesignSystem
 /// └── VStack
 ///     ├── CalendarHeaderView (기간 네비게이션 관련)
 ///     └── TabView (페이징 가능한 기간별 캘린더)
-///         └── PeriodView[] (SalaryBudget 개수만큼 있음)
+///         └── PeriodPageView[] (SalaryBudget 개수만큼 있음)
 ///             └── CalendarGridView (달력 그리드)
 ///                 ├── WeekdayHeaderView (요일 헤더)
 ///                 └── CalendarCell[] (DailyBudget 개수만큼 있음)
@@ -36,9 +36,9 @@ import DesignSystem
 /// - 이전/다음 기간으로 이동할 수 있는 네비게이션 버튼 제공
 /// - 기간 이동 가능 여부에 따라 버튼 활성화/비활성화
 ///
-/// ### 2. TabView with PeriodView
+/// ### 2. TabView with PeriodPageView
 /// - 여러 예산 기간을 페이징 방식으로 탐색 가능
-/// - 각 PeriodView는 한 예산 기간의 전체 캘린더를 표시
+/// - 각 PeriodPageView는 한 SalaryBudget 캘린더 페이지를 의미
 /// - 스와이프 제스처로 기간 간 이동 지원
 ///
 /// ### 3. CalendarGridView
@@ -147,7 +147,7 @@ struct CalendarView: View {
           }
         )) {
           ForEach(0..<viewModel.state.periodsCount, id: \.self) { index in
-            PeriodView(
+            PeriodPageView(
               viewModel: viewModel
             )
             .tag(index)
@@ -194,51 +194,6 @@ struct CalendarView: View {
         viewModel.send(.movePreviousPeriod)
       }
     }
-  }
-}
-
-// MARK: - Period View
-private struct PeriodView: View {
-  let viewModel: CalendarViewModel
-  
-  var body: some View {
-    ScrollView(showsIndicators: false) {
-      CalendarGridView(
-        daysInPeriod: createDaysInPeriod(
-          start: viewModel.state.currentPeriod.start,
-          end: viewModel.state.currentPeriod.end
-        ),
-        selectedDate: viewModel.state.selectedDate,
-        dayInfos: viewModel.state.dayInfos,
-        onDateSelected: { date in
-          withAnimation {
-            viewModel.send(.onDateSelected(date))
-          }
-        }
-      )
-      .padding(.top, 18)
-      .padding(.horizontal, 14)
-    }
-  }
-  
-  /// startDate와 endDate를 받아 캘린더에 표시할 기간을 생성합니다.
-  private func createDaysInPeriod(start: Date, end: Date) -> [Date?] {
-    let calendar = Calendar.current
-    var dates: [Date?] = []
-    
-    let startWeekday = calendar.component(.weekday, from: start) - 1
-    dates += Array(repeating: nil as Date?, count: startWeekday)
-    
-    var currentDate = start
-    while currentDate <= end {
-      dates.append(currentDate)
-      currentDate = calendar.date(byAdding: .day, value: 1, to: currentDate)!
-    }
-    
-    let remainingDays = (7 - (dates.count % 7)) % 7
-    dates += Array(repeating: nil as Date?, count: remainingDays)
-    
-    return dates
   }
 }
